Import of HIO/Unicode-Japanese-0.46 from CPAN.
gitpan-cpan-distribution: Unicode-Japanese
gitpan-cpan-version:      0.46
gitpan-cpan-path:         HIO/Unicode-Japanese-0.46.tar.gz
gitpan-cpan-author:       HIO
gitpan-cpan-maturity:     released

+++ wordsize.c
@@ -3,10 +3,10 @@
 
 int main()
 {
-  printf( "short = %d\n", sizeof(short) );
-  printf( "int = %d\n",   sizeof(int)   );
-  printf( "long = %d\n",  sizeof(long)  );
-  printf( "void* = %d\n", sizeof(void*) );
+  printf( "short = %ld\n", (long)sizeof(short) );
+  printf( "int = %ld\n",   (long)sizeof(int)   );
+  printf( "long = %ld\n",  (long)sizeof(long)  );
+  printf( "void* = %ld\n", (long)sizeof(void*) );
   return 0;
 }
